Request scoped construction in FastApi (#181)
Adds a new container ContextContainer which wraps other containers and adds an __enter__ and __exit__. Whenever specified the classes are requested the context managers are used to load them and the __exit__s of the context managers added to the ContextContainer's exit. The FastAPI integration makes use of one of these containers.